asteroid: Center sprite with offset
Center the sprite with an offset rather than changing its position. That
way scaling the sprite and the collision polygon by the same values
maintains the polygon points at the asteroid vertices. I believe this
also improves the center of mass handling as the sprite's origin is
roughly its center of mass.

https://phabricator.endlessm.com/T35418