ci: Add workflow to deploy to GitHub Pages
While it is not necessarily useful to have this game in its default state
published to the web, this workflow can be used by a learner who has customised
the game in their own repo.

The workflow is designed to fail gracefully if GitHub Pages is not available
(which is typically the case in a private repo owned by a free account) or not
configured correctly.

It also only runs on releases or when manually triggered, to avoid running it
each time a learner pushes or merges to main.

https://phabricator.endlessm.com/T35699