ci: Cancel running workflows if retriggered
For rationale, see
<https://turso.tech/blog/simple-trick-to-save-environment-and-money-when-using-github-actions>.

For the linting & formatting action, the concurrency group is set to
contain the PR (if any) and ref otherwise, on the basis that only the
latest push to a PR or branch is interesting.

For the GitHub Pages and Godot Asset Library actions, each run
overwrites the result of any previous run, so the concurrency group is
just the workflow name.