HUD: Suppress font_size narrowing conversion warning
    W 0:00:00:396   Narrowing conversion (float is converted to int and loses precision).
      <GDScript Error>NARROWING_CONVERSION
      <GDScript Source>hud.gd:41

This is because the exported font_size property is a float, but Label's
font_size theme property is an int.

I guess the reason that the exported property is a float is that
@export_range for a float gives a slider, while @export_range for an int
gives a spinbox.

Suppress the warning by adding an explicit roundi() call to convert from
float to int. Adjust the exported property's step to be 1 rather than
0.1 given that fractional adjustments will be ignored anyway.