basic_paddle: Don't await ready in setters
This is unnecessary because we reapply the properties in _ready() by
invoking the setters again.

It's also actively harmful: in web exports, it causes the assigment of
non-default values of the `player` property to be lost, meaning that
both paddles in the main scene are controlled by the arrow keys
(Player.RIGHT). I have to say that I don't fully understand why it
happens, but this change does fix it!

Fixes https://github.com/endlessm/moddable-pong/issues/11