basic_ball: Remove unused DEFAULT_BALL_SIZE constant
This was added in commit a7bf4ad89e961ed48468ed1ff9ba02e8409b7991. The
linter doesn't like its position in the file (the style guide says that
constants should come before variables) but it's also unused, so can be
removed. The constant that is actually used is _INITIAL_RADIUS.

https://docs.godotengine.org/en/stable/tutorials/scripting/gdscript/gdscript_styleguide.html#code-order