dependabot: Group updates for actions/* namespace
It is common for GitHub to release new versions of all their actions in
near-lockstep, causing a flood of PRs in each of our projects.

We can't help the fact we have multiple projects, but having read the
documentation it looks like we can at least group these updates together.